+++ app/src/main/java/com/example/simplychatgptapiconsumer/common/component/ResultText.kt
