+++ app/src/main/java/com/example/simplychatgptapiconsumer/main/intent/MainViewSideEffect.kt
@@ -4,5 +4,5 @@ import com.example.simplychatgptapiconsumer.common.mvicomponents.MviSideEffect
 
 sealed class MainViewSideEffect() : MviSideEffect {
 
-    data class ShowSnackBar(val text: String) : MainViewSideEffect()
+    data class ShowSnackBar(val snackBarMessage: String) : MainViewSideEffect()
 }
