+++ app/src/main/java/com/example/simplychatgptapiconsumer/main/mvimodel/MainViewState.kt
@@ -6,5 +6,6 @@ import com.example.simplychatgptapiconsumer.main.mvimodel.ResponseType.IDLE
 data class MainViewState(
     val chatResponse: String = "",
     val triviaSubject: String = "",
+    val isTriviaSubjectTooLong: Boolean = false,
     override val type: ResponseType = IDLE
 ) : MviState()
